+++ src/business/calendar/ui/web/index.ts
@@ -1,2 +1,0 @@
-export * from '../../../../common/components';
-export * from './pages';