+++ src/business/calendar/ui/web/pages/index.ts
@@ -1,2 +1,0 @@
-export { CalendarPage } from '../../../pages/CalendarPage';
-export { default as EventDetailPage } from '../../../pages/EventDetailPage';