+++ src/business/index.ts
@@ -4,7 +4,6 @@ export * as portfolio from './portfolio';
 export * as navigation from './navigation';
 export * as testField from './testField';
 export * as mmd from './mmd';
-export * as calendar from './calendar';
 export * as testYourself from './testYourself';
 export * as music from './music';
 export * as mikuFusionGame from './mikuFusionGame';